fix: quicklink service init (#830)
